Feat: Removed other class moved into library
