--- src/main/java/it/pagopa/transactions/domain/TransactionAmount.java
@@ -1,5 +1,0 @@
-package it.pagopa.transactions.domain;
-import it.pagopa.ecommerce.commons.annotations.ValueObject;
-
-@ValueObject
-public record TransactionAmount(int value) {}
