--- src/main/java/it/pagopa/transactions/domain/TransactionDescription.java
@@ -1,6 +1,0 @@
-package it.pagopa.transactions.domain;
-
-import it.pagopa.ecommerce.commons.annotations.ValueObject;
-
-@ValueObject
-public record TransactionDescription(String value) {}
