--- src/main/java/it/pagopa/transactions/domain/TransactionId.java
@@ -1,6 +1,0 @@
-package it.pagopa.transactions.domain;
-import it.pagopa.ecommerce.commons.annotations.ValueObject;
-import java.util.UUID;
-
-@ValueObject
-public record TransactionId(UUID value) {}
